Add files via upload
This is the VisualStudio project that I created. Right now, it is giving 2x LNK2019 Errors with a LNK1120 Error saying that there are 2 unresolved externals in main.obj. I don't know how to fix this, but I am working on figuring it out and will update as soon as I fix it.